[RN] rearrange display-name files for later refactor
