+++ react/features/conference/components/native/styles.js
@@ -22,20 +22,6 @@ export default createStyleSheet({
         flex: 1
     }),
 
-    displayNameBackdrop: {
-        alignSelf: 'center',
-        backgroundColor: 'rgba(28, 32, 37, 0.6)',
-        borderRadius: 4,
-        margin: 16,
-        paddingHorizontal: 16,
-        paddingVertical: 4
-    },
-
-    displayNameText: {
-        color: ColorPalette.white,
-        fontSize: 14
-    },
-
     gradient: {
         position: 'absolute',
         top: 0,
