+++ react/features/display-name/components/index.native.js
@@ -1,0 +1,3 @@
+// @flow
+
+export * from './native';
