+++ react/features/display-name/components/index.web.js
@@ -1,0 +1,3 @@
+// @flow
+
+export * from './web';
