+++ react/features/display-name/components/native/index.js
@@ -1,0 +1,4 @@
+// @flow
+
+export { default as DisplayNameLabel } from './DisplayNameLabel';
+export { default as DisplayNamePrompt } from './DisplayNamePrompt';
