+++ react/features/display-name/components/native/styles.js
@@ -1,0 +1,19 @@
+// @flow
+
+import { ColorPalette } from '../../../base/styles';
+
+export default {
+    displayNameBackdrop: {
+        alignSelf: 'center',
+        backgroundColor: 'rgba(28, 32, 37, 0.6)',
+        borderRadius: 4,
+        margin: 16,
+        paddingHorizontal: 16,
+        paddingVertical: 4
+    },
+
+    displayNameText: {
+        color: ColorPalette.white,
+        fontSize: 14
+    }
+};
