+++ react/features/display-name/components/web/index.js
@@ -1,2 +1,4 @@
+// @flow
+
 export { default as DisplayName } from './DisplayName';
 export { default as DisplayNamePrompt } from './DisplayNamePrompt';
