+++ react/features/display-name/index.js
@@ -1,3 +1,5 @@
+// @flow
+
 export * from './actions';
 export * from './components';
 export * from './functions';
